Develop

+++ src/main/java/com/plim/plimserver/domain/allergy/repository/AllergyCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.plim.plimserver.domain.allergy.domain.AllergyCategory;
 
-public interface AllergyCategoryRepository extends JpaRepository<AllergyCategory, Long>{
+public interface AllergyCategoryRepository extends JpaRepository<AllergyCategory, Integer>{
 
 }
