+++ src/main/java/com/plim/plimserver/domain/allergy/repository/FoodAllergyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.plim.plimserver.domain.allergy.domain.FoodAllergy;
 
-public interface FoodAllergyRepository extends JpaRepository<FoodAllergy, Long> {
+public interface FoodAllergyRepository extends JpaRepository<FoodAllergy, Integer> {
 
 }
